Targeting Android 13 (compileSdkVersion 33, targetSdkVersion 33) Removed deprecated LocalBroadcastManager and ViewPager

+++ app/src/main/java/nl/lxtreme/binutils/hex/IntelHexReader.java
@@ -148,7 +148,7 @@ public class IntelHexReader extends AbstractReader
         else
         {
           // Decrease the number of hex-bytes we've got to read...
-          this.checksum.add( ( byte )dataByte );
+          this.checksum.add(dataByte);
 
           this.dataLength--;
           this.address++;
