+++ app/src/main/java/nl/lxtreme/binutils/hex/SRecordReader.java
@@ -136,7 +136,7 @@ public class SRecordReader extends AbstractReader
         else
         {
           // Decrease the number of hex-bytes we've got to read...
-          this.checksum.add( ( byte )dataByte );
+          this.checksum.add(dataByte);
 
           this.dataLength--;
           this.address++;
